New chapter "Hyper-converged Infrastructure" with Proxmox ceph server
The new chapter "Hyper-converged Infrastructure" is added. It contains a general statement about it and
as subchapter "Ceph server in Proxmox VE Cluster"

The latter is derived from the corresponding Wiki article. Since the Wiki article contains a lot of
additional best practice information it is not obsolete yet. A combined source for both Wiki and Administration Guide
will be made later.

Signed-off-by: Friedrich Ramberger <f.ramberger@proxmox.com>