pveceph: followup: refactor advantage list
remove duplicated 'Easy management' point.
Remove 'no signle point of failure' as it's very easy to setup a
ceph cluster with lot's of single point of failures, ceph allows to
avoid them, but does not magically do so - so do not lull an unknown
user into false hope.
Use {pve} makro and rewor heading to a more "what's cool on ceph
_with_ pve" one.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>