api-viewer: use our extjs package
The loading of an external hosted ExtJS is from a time where we did
not have a extra ExtJS package as it still was in the pve-manager
one, which was not desired to be installed on a non-pve machine.
Add a dependency on libjs-extjs and a link from api-viewer/extjs to
the installation directory of the ExtJS package, this allows to reuse
ExtJS on PVE node hosted docs and on the hosted docs from proxmox.com
at the same time.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>