Fix 1891: Add zsh completion generator
This adds the generator needed for the autogeneration of zsh completion scripts
analogous to the generator creating the bash completion scripts.
In order to generate zsh completion scripts for other packages, a package
including this patch as well as the package in pve-common with the corresponding
patch have to be installed on the build system.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>