ceph: move ceph docs to own top level chapter
This change should help to make the meanwhile extensive documentation
about ceph more visible, therefore ceph now has it's own chapter which
is sorted more reasonable after the storage chapter. The formerly known
"hyper converged infrastructure" chapter is now integrated in the intro
chapter, as it only contains some general information about this topic,
which is not specific for ceph.

Signed-off-by: Tim Marx <t.marx@proxmox.com>