mediawiki: bring doc includer again in line with the HTMLets plug it bases on
We probably always wanted to use a real HTML comment as "tag" here,
but it seems the '!' was forgotten. With older mediawiki it still
worked, but 1.31 got stricter and thus it was converted to HTML
entities.

Bring the plugin again in line with the HTMLets MediaWiki
extension[0] it based on in the first place.

[0]: https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/extensions/HTMLets/+/11e5ef1ea2820319458dc67174ca76d6e00b10cc/HTMLets.php#140

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>