switch opt-out NOVIEW to opt-in PVE_DOC_INSTANTVIEW
As we had to set NOVIEW _everywhere_, which is a good sign that
opt-in is much better suited than opt out.

Especially as one can point the browser or man to the resulting build
documentation html file or manpage once and then hit just reload after
a re-build, which is almost zero effort.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>