cert-management: mention symlinks in /etc/pve
The warning to not replace the cluster-certificates in '/etc/pve/local' can
be misleading and let users think that '/etc/pve/nodes/NODENAME/pve-ssl.pem'
(and .key) are the files they should replace with a LE/externally signed
certificate.

Explicitly mentioning that '/etc/pve/local' is a symlink to
'/etc/pve/nodes/NODENAME' should make the warning more clear.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>