sysboot: add a section on how to determine the used bootloader
Besides observing the boot process, `efibootmgr` seems to be the safest
bet. It should not show anything if booted in legacy BIOS mode. If
booted in UEFI mode, the boot entries should show which bootloader is
present.

Having both present should not happen in a regular use case AFAICT and
would need more consideration, or better, a reboot.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>