old repos: use /pve suffix, which has better compatibility
We switched configuring /etc/apt/auth.conf to use /debian/pve as path
instead of just the old /debian for better compatibility with PVE and
PBS getting co-installed on the same host.

To match that stricter server definition the sources.list for
pve-enterprise must contain the /pve at the end. Otherwise there is
no match and updating via the pve-enterprise repository doesn't work.

Signed-off-by: Mira Limbeck <m.limbeck@proxmox.com>
 [ Thomas: adapted commit message and older repos as 6.x too ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>