network: remove ifup ifdown as it won't reattach guests
Mentioning 'ifup' and 'ifdown' can lead to users running into situations
where they take down one interface, for example an vmbrX. In this
situation the connection to the guests will be lost and not
reestablished automatically if the interface is brought up again with
'ifup'. The better way to apply network changes is to run 'ifreload -a'.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>