network: rework and reorder reload/apply section
Order the live-reload method first & mention that it's recommended.

Mention that older PVE installations may also need to install
ifupdown2 manually, not just vanilla-Debian based ones.

Demote the `ifreload -a` into an inline mention, and emphasize that
it really only is for manual direct edits by moving that to the front
of the sentence, to avoid that users think that ifreload -a is an
alternative to the apply changes button/api.

Explain, in short, how changes are applied on reboot.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>