network: rephrase corosync and bonds recommendations
I suspect that the old one seems to be related to multicast traffic and
LACP bonds.

The link in the comment is dead by now. It seems this is one occasion
where the internet actually forgets as I cannot find the actual message
of that mailing list thread anymore. Therefore I cannot say for sure
what the exact issue was. But it was introduced in commit
649098a64ecaffc7215ec0556e76787595b38e88 which unfortunately also
doesn't have more information.

Since with Corosync 3, unicast is used, that recommentation is probably
not accurate anymore. At least I am not aware of any issues with
Corosync on LACP bonds in recent years. Therefore, rather recommend to
configure Corosync on multiple networks instead of bonds to follow best
practice.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>