d/control: ensure docs are build with compatible asciidoc
As commit 5832a54 ("fix running a2x under new package structure") in
the upstream asciidoc-py repo broke overriding the backend with our
own config for the docbook target, which broke our thumbnail
attribute for paragraphs for the PDF output (HTML is done directly by
asciidoc, and there overriding the config still worked).

Reported upstream [0] and workarounded with the proposed fix applied
to an asciidoc package re-build that got uploaded to our devel repo
for bookworm, so ensure that builds are done with a new enough
version.

[0]: https://github.com/asciidoc-py/asciidoc-py/issues/266

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>