installation: note that TUI uses same code and use thumbnail attribute
The thumbnail attribute is our preferred way to include those
screenshots, but it requires to be inside a paragraph to be replaced, and
having an admonition directly below the attribute doesn't makes it
part of a paragraph.

So, add a bit of text, which was warranted anyway, allowing us to use
the thumbnail attr again.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>