calendar events: fix build for wiki target
This was actually just some POC that got committed by accident and
broke build to to getting a title where not expecting it as there
where two (relatively seen) top-level titles in the calendar events
page.

Fix that by declaring an actual level-0 heading (i.e., title) in the
wiki context.

This then should also fix a dangling link to the calendar-events wiki
page for the docs-in-wiki stuff, e.g., "calendar events" in backup
chapter [0] pointint to 404 [1], that Fiona reported a few weeks ago.
[0]: https://pve.proxmox.com/wiki/Backup_and_Restore#vzdump_jobs
[1]: https://pve.proxmox.com/wiki/Schedule_Format#chapter_calendar_events

Fixes: a90b75d ("package repos: add Ceph 18.2 Reef repos")
Reported-by: Lukas Wagner <l.wagner@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>