fix #5429: network: override device names: include Type=ether
Mention that the systemd link file should contain `Type=ether` in most
setup, to make sure it only applies to Ethernet devices and does not
ever apply to e.g. bridges or bonds which inherit the MAC address of
the Ethernet device. Mention that some setups may require other
options.

Reported in the forum [0] and in #5429 [1].

[0] https://forum.proxmox.com/threads/144557/post-656188
[1] https://bugzilla.proxmox.com/show_bug.cgi?id=5429

Fixes: 96c0261 ("fix #4847: network: extend section on interface naming scheme")
Signed-off-by: Friedrich Weber <f.weber@proxmox.com>