network: override device names: suggest running update-initramfs
The initramfs-tools hook /usr/share/initramfs-tools/hooks/udev copies
link files from /etc/systemd/network to the initramfs, where they take
effect in early userspace. If the link files in the initramfs diverge
from the link files in the rootfs, this can lead to confusing
behavior, as reported in enterprise support. For instance:

- If an interface matches link files both in the initramfs and the
  rootfs, it will be renamed twice during boot.
- A leftover link file in the initramfs renaming an interface A to a
  new name X may prevent a link file in the rootfs from renaming a
  different interface B to the same name X (it will fail with "File
  exists").

To avoid this confusion, mention the link files are copied to the
initramfs, and suggest updating the initramfs after making changes to
the link files.

Suggested-by: Hannes Laimer <h.laimer@proxmox.com>
Signed-off-by: Friedrich Weber <f.weber@proxmox.com>