passthrough: viommu: replace host requirement with reference to pcie passthrough
Enabling IOMMU on the host is not a requirement for vIOMMU. It is only
a requirement for passthrough. Add a sentence to clarify the need for
a configured PCI(e) passthrough on the host for passthrough to nested
VMs.

While at it also clarify that this does not only applies to PCI but
also PCIe here.

Suggested-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Markus Frank <m.frank@proxmox.com>
 [ TL: mention all changes and avoid overly long line with link ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>