qm, vzdump: add notes regarding backups of windows and VSS
VSS in windows guests with additional backup software in the guest can
cause problems. By now the guest agent does support to use a different
VSS option. This way, one can switch to one that does not interfere with
other backup solutions.

Mention it in the docs and link to the wiki article that explains it.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Reviewed-by: Alexander Zeidler <a.zeidler@proxmox.com>