virtual machines: rework & extend PVE machine version section
Make the PVE machines a sub point of the overall machine version, as
they are not really different, but only extend the mechanism we
already used by something we can directly control. I tried to extend
the previously existing machine version to also describe that we use
that version for deciding what to enabled/change on the PVE side,
i.e., not just what QEMU exposes, and also reduce some verbosity with
regard to implementation details.

While at it, I replaced the first-person plural (we) use with third
person indicative mood as documented in our style guide [0].

[0]: https://pve.proxmox.com/wiki/Technical_Writing_Style_Guide#Person_and_Mood

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>