pveceph: update cluster shutdown procedure
It seems that the shutdown procedure which we had documented (and many
others as well) can potentially cause problems when the cluster starts
up again.

There is a blog article from croit explaining it in detail [0] if you
are interested.

The shorter explanation is that when unsetting 'pause' it might take
half an hour to get out of that state. Unsetting nodown also could take
a long time. During all this OSDs and MONs might consume a lot of CPU
resources, increasing the CPU load on the host considerbly.

nobackfill and norecover are not needed, but also not harmful.

Archeology revealed that the procedure with all the other flags (which
this commit removes), originated in the RedHat documentation, but was
never part of Ceph's shutdown procedure which is tested by the Ceph
team.

[0] https://web.archive.org/web/20250624082830/https://www.croit.io/blog/how-not-to-shut-down-a-ceph-cluster

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Link: https://lore.proxmox.com/20250624085306.3224755-1-a.lauterer@proxmox.com