notifications: adapt to latest notification system UI changes
The latest updates to the backup-job UI completely drop the term
"Notification System" from the UI, instead we now use "Global
notification settings", which should be hopefully a bit clearer to users
with regards to what this actually means.

Furthermore, the 'auto' notification mode is not exposed anymore in the
UI, hence we also drop it from the documentation to avoid confusion. It
is still listed in the backup job option reference section, for any user
who might still be using it.

Some of the touched sections were slightly rephrased to improve clarity.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20250725082046.51199-1-l.wagner@proxmox.com