remove GlusterFS plugin documentation
Mirroring pve-storage commit 7669a99 ("drop support for using
GlusterFS directly"). Native GlusterFS support will not be available
for PVE 9, one still can manually mount and use such a storage as
directory storage though, or keep using PVE 8 until EOL and use that
time for migrating to e.g. Ceph or the like.

The main reason is that upstream has stopped development and various
integrations are planned to go away soon, like the QEMU one. That's
why we cannot promote it as stable storage for PVE 9.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>