mediawiki: return error page instead of exception on bad name
Doing a die here does not help us, it will only spam the apache error
log on the server and is user controlled, so could be abused to hide
other errors or just fill up the error log until disks are full or the
like.

additionally limit the usage of points even further, i.e. require that
it's followed by a alphanumeric or hyphen char. While we did not allow
a slash, thus no breakout was possible, it's still not necessary to
allow ".." here, so better stay safe than sorry, like if this gets
ever extended for sub-folder support.

Rather just return a error message to the client.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>