qm: pci passthrough: correct note about using 'softdep'
In most cases with passthrough, the vfio-pci driver is preferred to what
the kernel would normally load (e.g. amdgpu/nouveau), so 'softdep' can
be used to make sure 'vfio-pci' is loaded before the actual module.

This is expressed in our docs already with a sensible example, but the
text is wrong, since it says 'load the gpu modules *before* loading
vfio-pci'.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251219083102.623503-1-d.csapak@proxmox.com
[FE: "use vfio-pci before gpu" rather than "gpu after vfio-pci"]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>